Add Contract Tests
I am writing a blog post that illustrates Liskov's Substitution
Principle with Contract Tests. I added contract tests for Java's Set
class and a few of it's implementations as a means to that end.

I have not finished the contract tests completely, but I have a stable
point that compiles, runs, and passes the junit tests.

I have to do a better job at documenting the code, though. I believe
that the next commit will contain more documentation.