Restructure Directory For Human Comprehension
I used the test directory as a global dumping ground. I really should
have organized it better from the start.

I have rectified that problem by creating a directory structure that
reflects the purpose of each file. For example, under the directory
set you will find a the SetContractTest abstract class, and under the
directory impl you will find the implementations of that class.

I believe that this will help a reader better comprehend the design and
purpose of the program.