Add CountIntegers Program
I added a program that will count the number of unique integers in an
array of integers. I wanted to have a simple program that uses the set
interface to accomplish some goals. I will use this program as an
example of the liskov substitution principle. The particular type of set
does not matter for program correctness. The set simply has to obey the
rules of a set.